refactor take most of ProcessWrapper out into a custom hook, move redunant markup to other files
