+++ src/altinn-app-frontend/src/common/hooks/index.ts
@@ -16,6 +16,7 @@ export const usePrevious = (value: any) => {
   });
   return ref.current;
 };
+export { useProcess } from './useProcess';
 export { useAppSelector } from './useAppSelector';
 export { useAppDispatch } from './useAppDispatch';
 export { useInstanceIdParams } from './useInstanceIdParams';
