+++ src/altinn-app-frontend/src/common/hooks/useApiErrorCheck.ts
@@ -1,0 +1,8 @@
+import { useAppSelector } from 'src/common/hooks/useAppSelector';
+import { makeGetHasErrorsSelector } from 'src/selectors/getErrors';
+
+export function useApiErrorCheck() {
+  const hasErrorSelector = makeGetHasErrorsSelector();
+  const hasApiErrors = useAppSelector(hasErrorSelector);
+  return { hasApiErrors };
+}
